requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/whl/windows/cpu/avx/stable.html

python-dateutil==2.9.0.post0
selenium==4.31.0
webdriver-manager==4.0.2

paddlepaddle==2.6.1          # CPU-only
paddleocr==2.7.0.3
opencv-python-headless==4.10.0.82
scikit-learn==1.4.2
pillow==10.3.0
numpy==1.24.4